protected the CRUD routes

+++ backend/src/controllers/goalControllers.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 //@Desc:  Get goals
 //@route: GET /api/goals
 //@access: Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -20,8 +21,8 @@ const setGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Kindly add text in the goals field");
   }
-  console.log("we here");
-  const goal = await Goal.create({ text });
+
+  const goal = await Goal.create({ text, user: req.user.id });
 
   res.status(200).json(goal);
 });
@@ -36,6 +37,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("The resource does not exist.");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(id, req.body, {
@@ -55,6 +70,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("The resource does not exist.");
   }
+
+ const user = await User.findById(req.user.id);
+
+ //check for user
+ if (!user) {
+   res.status(401);
+   throw new Error("User not found");
+ }
+
+ //make sure the logged in user matches the goal user
+ if (goal.user.toString !== user.id) {
+   res.status(401);
+   throw new Error("User not authorized");
+ }
 
   const deletedGoal = await Goal.findByIdAndDelete(id);
   res.status(200).json(deleteGoal);
